mistral_setup/requirements.txt
<<<<<<< HEAD
supabase
flask
flask_cors
flask_sqlalchemy
python-dotenv
requests
=======
streamlit==1.35.0
pandas==2.2.2
openpyxl==3.1.2
email-validator==2.2.0
openpyxl==3.1.2
>>>>>>> 0ab7eedc0051802a114a7c3a324c57021f6175bb
